Refactor routes into APIRouter modules
