+++ app/dependencies.py
@@ -1,0 +1,18 @@
+from fastapi import Depends, HTTPException, Request, status
+from fastapi.templating import Jinja2Templates
+from sqlalchemy.orm import Session
+
+from .database import get_db
+from . import models
+
+templates = Jinja2Templates(directory="templates")
+
+
+def get_current_user(request: Request, db: Session = Depends(get_db)):
+    user_id = request.session.get("user_id")
+    if user_id is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    user = db.query(models.User).filter(models.User.id == user_id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
+    return user
